+++ idp-commons/src/test/java/de/gematik/idp/ClientUtilitiesTest.java
@@ -16,9 +16,12 @@
 
 package de.gematik.idp;
 
+import static de.gematik.idp.field.ClientUtilities.generateCodeVerifier;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import de.gematik.idp.field.ClientUtilities;
+import java.util.Arrays;
+import org.apache.commons.lang3.ArrayUtils;
 import org.junit.jupiter.api.Test;
 
 class ClientUtilitiesTest {
@@ -28,7 +31,7 @@ class ClientUtilitiesTest {
 
   @Test
   void generateCodeChallengeFromVerifier() {
-    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String codeVerifier = generateCodeVerifier();
 
     final String codeChallenge = ClientUtilities.generateCodeChallenge(codeVerifier);
 
@@ -36,5 +39,29 @@ class ClientUtilitiesTest {
         .matches(BASE64_URL_REGEX)
         .isEqualTo(ClientUtilities.generateCodeChallenge(codeVerifier))
         .hasSize(SHA256_AS_B64_LENGTH);
+  }
+
+  /*
+   * https://datatracker.ietf.org/doc/rfc7636/
+   * Appendix B.  Example for the S256 code_challenge_method
+   */
+  @Test
+  void getCodeVerifierAndChallengeIetf() {
+    final int[] ietfExampleOctets = {
+      116, 24, 223, 180, 151, 153, 224, 37, 79, 250, 96, 125, 216, 173, 187, 186, 22, 212, 37, 77,
+      105, 214, 191, 240, 91, 88, 5, 88, 83, 132, 141, 121
+    };
+    final String codeVerifierIetf = "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk";
+    final String codeChallengeIetf = "E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM";
+
+    final Byte[] objByteArray =
+        Arrays.stream(ietfExampleOctets).boxed().map(Integer::byteValue).toArray(Byte[]::new);
+    final byte[] primitiveByteArray = ArrayUtils.toPrimitive(objByteArray);
+
+    final String codeVerifier = generateCodeVerifier(primitiveByteArray);
+    assertThat(codeVerifier).isEqualTo(codeVerifierIetf);
+
+    final String codeChallenge = ClientUtilities.generateCodeChallenge(codeVerifier);
+    assertThat(codeChallenge).isEqualTo(codeChallengeIetf);
   }
 }